Initial commit
This is not quite the version of Big Red that was submitted for marking in
December. Here are the big changes:

* The namespace is now dk.itu rather than net.ybother. (Land grab!)
* The "metaclass" terminology has been replaced with "control", which it
  should have been all along.
* Various Actions have been consolidated into the new WizardAction.
* WorkbenchWizard has been replaced with SubWizard (and, as a result, its
  interfaces no longer require a handle to the workbench).
* The miscellaneous grab-bag of rubbish that was the big_red.Utility class
  has been moved to the big_red.util package and is being split up into
  more coherent sub-classes.

This commit also contains some of the initial work to make the model match
the formal bigraphs model more closely (ports, for instance, are being
promoted to first-class objects rather than just being string properties
of controls), although this isn't complete yet.