[Bigraph] Renamed root element to "brs", and removed "ports" tag
** BREAKS FILE-FORMAT BACKWARDS COMPATIBILITY **

As well as breaking the file format, this also probably breaks the editor
in important ways - the notion of port compatibility has been removed from
the file format, but not from the editor's behaviour and expectations. Oh,
well; one can afford to be more reckless when version control is available!