[AbstractPart] Changed AbstractPart.getModelChildren to use ILayoutable
This is an important step: not only does the editor still work (which is
always a plus), but now things which aren't Things (helpful names...) can
appear on the bigraph.

(On this note, AbstractPart is a misleading name - it's essentially a
ThingPart. This will be fixed later, either by renaming it or by making it
more generic.)