[ILayoutableRelayoutCommand] Added a parent layout size check
As you may have been able to guess, the check is taken entirely from
ILayoutableAddCommand. It's not quite good enough (in both contexts, it
should also check for overlaps with other children), but it's a start! This
commit also changes the public methods of the class to take Objects, since
that seems to work nicely for Commands