[Control] Reorganised the public API to add segments
This isn't all it does! Internally, Controls now keep track of Points as an
array of actual Point objects with a null parent, which is much tidier than
the old "lots of separate" arrays solution. Some code which assumed that
only an offset (an integer offset, no less!) was required to position a
Port has also been removed