[BigraphXMLImport] Fixed the importing of outer names
Edges are no longer created on-demand; instead, if a Point references a
Link that doesn't exist yet, the details are stowed away for future use.
When that Link is eventually created, all of those pending connections are
created along with it

(This commit also makes a minor change which makes the code much safer: the
process method now gives newly-created objects parents before it does
anything else. This means that methods which travel up the hierarchy, like
getBigraph, won't crash.)