[ILayoutable, LayoutableModelObject] Turned setParent into a protected method of LayoutableModelObject
The intention here is that Container.addChild and Container.removeChild
should be the only public methods which call setParent. It's important that
x.getParent() == c implies c.hasChild(x), but before this change there was
no way to ensure this condition!

This has lots of knock-on effects, though. Container needs to be able to
call setParent on its children, which in turn means that those children
need to become LayoutableModelObjects rather than just ILayoutables; this
has touched an awful lot of code throughout Big Red.

On the other hand, this isn't too much of a problem; the implementations in
LayoutableModelObject are reasonably universal, and making the basic
renderable object into a class rather than an interface will make it easier
to extend the model without having to worry about lots of Big Red-specific
minutiae