[ReactionRuleXMLExport] Added support for exporting more kinds of changes
Bad news: GEF represents property sheet changes by sneaking completely
opaque commands onto the CommandStack, so it's impossible to tell when
objects are renamed. The rule editor is entirely dependent on being able to
introspect the command stack, and so this will probably force me to remove
the property sheet altogether; thanks, Eclipse!