[SimulationSpec] Added a new class for simulation specs
Those of you who remember commit cdde28cdf5e7e258295db3f9eb8a61b6eddbee5b
will remember that simulation specs -- groups of rules with an accompanying
model and a signature, ready to be given to a tool -- have been planned for
a long time. Now that the first prototype of the reaction rule editor is
almost ready, it's finally time to start adding them!