[BigraphCanvas] Restored the BigraphCanvas class
Commit d014bf23991f29140b4ea37b1ce60ea5b6d48e16 removed this class because
the checkpoint system for which it had been designed was about to be
removed. However, it turns out to have another, even better use: allowing
external plugins to display bigraphs in the Big Red style. To this end,
this commit also adds a new, exported "big_red.editors.utilities" package